% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{.start_profiling_step}
\alias{.start_profiling_step}
\title{Begin profiling a step}
\usage{
.start_profiling_step(step_name)
}
\arguments{
\item{step_name}{Unique label for the step.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Internally called by \code{profile_code()} but can be used manually to start
timing an arbitrary code segment. Must be paired with
\code{.end_profiling_step()} using the same \code{step_name}.
}
\seealso{
\code{.end_profiling_step}, \code{profile_code}
}
