% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{profile_code}
\alias{profile_code}
\title{Profile a block of code}
\usage{
profile_code(step_name, expr)
}
\arguments{
\item{step_name}{A short label for the profiled block.}

\item{expr}{A block of R code enclosed in \code{{}}.}
}
\value{
Result of evaluating \code{expr}.
}
\description{
Convenience wrapper that automatically calls \code{.start_profiling_step()} and
\code{.end_profiling_step()} around \code{expr}.
}
